[1.0.0-SNAPSHOT] Full package refactoring groupId refactoring build.gradle updated and improved
