+++ README.md
@@ -5,23 +5,21 @@
 [![Coverage](https://sonarcloud.io/api/project_badges/measure?project=GoodforGod_uri-builder&metric=coverage)](https://sonarcloud.io/dashboard?id=GoodforGod_uri-builder)
 [![Maintainability Rating](https://sonarcloud.io/api/project_badges/measure?project=GoodforGod_uri-builder&metric=sqale_rating)](https://sonarcloud.io/dashboard?id=GoodforGod_uri-builder)
 
-Little library with helper for building URI in Java easier.
-
-Library supposed to move URIBuilder functionality out of Micronaut, so it can be reusable in non Micronaut applications.
+Little library to help build URI in Java easier.
 
 ## Dependency :rocket:
 
-[**Gradle**](https://mvnrepository.com/artifact/com.github.goodforgod/uri-builder)
+[**Gradle**](https://mvnrepository.com/artifact/io.goodforgod/uri-builder)
 ```groovy
 dependencies {
-    compile 'com.github.goodforgod:uri-builder:1.0.0'
+    implementation "io.goodforgod:uri-builder:1.0.0"
 }
 ```
 
-[**Maven**](https://mvnrepository.com/artifact/com.github.goodforgod/uri-builder)
+[**Maven**](https://mvnrepository.com/artifact/io.goodforgod/uri-builder)
 ```xml
 <dependency>
-    <groupId>com.github.goodforgod</groupId>
+    <groupId>io.goodforgod</groupId>
     <artifactId>uri-builder</artifactId>
     <version>1.0.0</version>
 </dependency>
