+++ build.gradle
@@ -5,7 +5,7 @@ plugins {
     id "java-library"
 
     id "org.sonarqube" version "3.3"
-    id "com.diffplug.spotless" version "5.11.0"
+    id "com.diffplug.spotless" version "5.14.3"
 }
 
 repositories {
@@ -17,14 +17,15 @@ repositories {
 group = groupId
 version = artifactVersion
 
-sourceCompatibility = 1.8
-targetCompatibility = 1.8
+sourceCompatibility = JavaVersion.VERSION_1_8
+targetCompatibility = JavaVersion.VERSION_1_8
 
 spotless {
     java {
-        encoding "UTF-8"
+        encoding("UTF-8")
+        importOrder()
         removeUnusedImports()
-        eclipse().configFile "${projectDir}/config/codestyle.xml"
+        eclipse().configFile("${projectDir}/config/codestyle.xml")
     }
 }
 
@@ -37,7 +38,7 @@ sonarqube {
 }
 
 dependencies {
-    implementation "org.jetbrains:annotations:20.1.0"
+    implementation "org.jetbrains:annotations:22.0.0"
 
     testImplementation "org.codehaus.groovy:groovy-all:3.0.8"
     testImplementation "org.spockframework:spock-core:2.0-groovy-3.0"
@@ -56,26 +57,15 @@ test {
     }
 }
 
-tasks.withType(JavaCompile) {
-    options.encoding = "UTF-8"
-    options.incremental = true
-    options.fork = true
-}
-
-java {
-    withJavadocJar()
-    withSourcesJar()
-}
-
 publishing {
     publications {
         mavenJava(MavenPublication) {
             from components.java
 
             pom {
-                name = "Name"
+                name = "URI Builder"
                 url = "https://github.com/GoodforGod/$artifactId"
-                description = "Description."
+                description = "Little library to help build URI in Java easier."
 
                 license {
                     name = "Apache License 2.0"
@@ -111,6 +101,16 @@ publishing {
     }
 }
 
+java {
+    withJavadocJar()
+    withSourcesJar()
+}
+
+tasks.withType(JavaCompile) {
+    options.encoding("UTF-8")
+    options.incremental(true)
+    options.fork = true
+}
 
 check.dependsOn jacocoTestReport
 jacocoTestReport {
@@ -120,16 +120,16 @@ jacocoTestReport {
     }
 }
 
-if (project.hasProperty("signing.keyId")) {
-    apply plugin: "signing"
-    signing {
-        sign publishing.publications.mavenJava
-    }
-}
-
 javadoc {
     options.encoding = "UTF-8"
     if (JavaVersion.current().isJava9Compatible()) {
         options.addBooleanOption("html5", true)
+    }
+}
+
+if (project.hasProperty("signing.keyId")) {
+    apply plugin: "signing"
+    signing {
+        sign publishing.publications.mavenJava
     }
 }
