+++ gradle.properties
@@ -1,6 +1,6 @@
-groupId=com.github.goodforgod
+groupId=io.goodforgod
 artifactId=uri-builder
-artifactVersion=1.0.1-SNAPSHOT
+artifactVersion=1.0.0-SNAPSHOT
 buildNumber=1
