+++ src/main/java/io/goodforgod/net/uri/URIBuilder.java
@@ -1,4 +1,4 @@
-package io.net.uri.builder;
+package io.goodforgod.net.uri;
 
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -10,7 +10,7 @@ import java.util.Objects;
 import java.util.stream.Stream;
 
 /**
- * Helper for building URI (Micronaut copycat of UriBuilder)
+ * Helper for building URI (Inspired by Micronaut UriBuilder)
  *
  * @author Anton Kurako (GoodforGod)
  * @since 21.08.2021
